move majorHelper to global util dir
